[ar] Make the crawler resilient to a problem with Arabic Sputnik News
For a handful of URLs in their sitemap, the Arabic edition of Sputnik
News returns HTTP redirects, which Python's urllib2 tries to follow.
However, the web site occasionally returns redirect locations which
cannot be UTF8-decoded, so Python's urllib2 runs into a Unicode
decoding error before we see the URL. We catch the dedoding error and
return to our caller a '400 Bad Request' error, so the affected pages
get ignored in the crawl.

Not sure if this is really a client error inside Python's urllib2,
or if the error is on the server side. In any case, this only affects
a very small number of URLs, so we don't care.